add cache singleton and changed int by long

+++ ApplicationCore/Repositories/IAddressRepository.cs
@@ -4,6 +4,6 @@ namespace ApplicationCore.Repositories
 {
     public interface IAddressRepository : IRepository<Address>
     {
-        ValueTask<Address> GetByIdAsync(int id, CancellationToken cancellationToken = default);
+        ValueTask<Address> GetByIdAsync(long id, CancellationToken cancellationToken = default);
     }
 }
