+++ ApplicationCore/Services/IContactService.cs
@@ -4,6 +4,6 @@ namespace ApplicationCore.Services
 {
     public interface IContactService : IBaseService<Contact>
     {
-        ValueTask<int> CreateEmployeeForCompany(int companyId, Contact contact, CancellationToken cancellationToken = default);
+        ValueTask<long> CreateEmployeeForCompany(long companyId, Contact contact, CancellationToken cancellationToken = default);
     }
 }
