+++ Infrastructure/Data/CompanyRepository.cs
@@ -40,7 +40,7 @@ namespace Infrastructure.Data
                 await _context.Companies.ToListAsync();
         }
 
-        public async ValueTask<Company> GetByIdAsync(int id, CancellationToken cancellationToken = default)
+        public async ValueTask<Company> GetByIdAsync(long id, CancellationToken cancellationToken = default)
         {
             return await _context.Companies
                 .Where(p => p.Id == id)
