+++ Infrastructure/Entities/Exceptions/ContactRoleNotFoundException.cs
@@ -2,7 +2,7 @@
 {
     internal class ContactRoleNotFoundException : NotFoundException
     {
-        public ContactRoleNotFoundException(int contactRoleId) : base($"The contact role with id: {contactRoleId} doesn't exist in the database.")
+        public ContactRoleNotFoundException(long contactRoleId) : base($"The contact role with id: {contactRoleId} doesn't exist in the database.")
         {
         }
     }
